+++ CHANGELOG.md
@@ -20,3 +20,4 @@ Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED.
     - `sound_event`: The Sound Event to play when a Player is damaged by `damage_types`.
     - `config_predicate`: An optional field, used to determine whether this can be used.
   - Removed the initial implementation, as this new data-driven implementation is far more stable.
+- Added `ShearsDispenseItemBehaviorApi`, used to register custom behaviors for Shears being used by a Dispenser.